% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{area_hres}
\alias{area_hres}
\title{Upstream drainage area raster}
\format{\code{area_hres} is supplied with the \code{urdn} package.}
\usage{
area_hres
}
\description{
Raster data of Mogi-Guaçú/Pardo river basin.
}
\keyword{datasets}
