% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortran-connection.R
\name{cotat_plus}
\alias{cotat_plus}
\title{Upscaling river drainage network}
\usage{
cotat_plus(dirh, areah, factor, area.thres, mufp, mascfile = "")
}
\arguments{
\item{dirh}{high resolution flow direction grid.}

\item{areah}{high resolution drainage area grid.}

\item{factor}{integer. Aggregation factor expressed as number of cells in
each direction (horizontally and vertically). For example, fact=2 will result
in a new Raster* object with 2*2=4 times fewer cells.}

\item{area.thres}{area threshold (set equal to the cell area).}

\item{mufp}{minimum upstream flow path (set as one fifth of the cell size).}

\item{mascfile}{mask file}
}
\value{
list object with 2 elements:
 \describe{
  \item{hires}{RasterLayer with location of cells outlet}
  \item{lowres}{RasterStack with flow direction, outlet row, and outlet col}
  }
}
\description{
Upscaling flow direction from fine- to coarse‐resolution river networks
}
\examples{
area_hres
dir_hres
flowd_coarse_grid <- cotat_plus(dirh = dir_hres,
                               areah = area_hres,
                               factor = 10,
                               area.thres = 1.0,
                               mufp = 0.02)
# outlet locations (high resolution)
flowd_coarse_grid[[1]]
# flow direction, outlet row, outlet col
flowd_coarse_grid[[2]]
plot(flowd_coarse_grid[[2]])
flow_dir <- raster(flowd_coarse_grid[[2]], 1)
flow_dir
}
