% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\docType{data}
\name{dir_hres}
\alias{dir_hres}
\title{Flow direction raster}
\format{\code{dir_hres} is supplied with the \code{urdn} package.}
\usage{
dir_hres
}
\description{
Raster data of Mogi-Guaçú/Pardo river basin.
}
\keyword{datasets}
