% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urdn-package.R
\docType{package}
\name{urdn-package}
\alias{urdn}
\alias{urdn-package}
\title{urdn: Upscaling River Drainage Networks}
\description{
Upscaling River Drainage Networks.
}
\author{
\strong{Maintainer}: Jonatan Tatsch \email{jdtatsch@gmail.com}

}
\keyword{internal}
