% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortran-connection.R
\name{cotat_plus}
\alias{cotat_plus}
\title{Upscaling river drainage network}
\source{
Fotran source code provided by Dr.
Adriano Paz (\url{https://sites.google.com/site/adrianorpaz2})
}
\usage{
cotat_plus(flowdir.fres, draina.fres, factor, area.thres, mufp, mascfile = "")
}
\arguments{
\item{flowdir.fres}{fine resolution flow direction raster.}

\item{draina.fres}{fine resolution drainage area raster.}

\item{factor}{integer. Aggregation factor expressed as number of cells in
each direction (horizontally and vertically). For example, fact=10 will result
in a new Raster* object with 10*10=100 times fewer cells.}

\item{area.thres}{area threshold (set equal to the cell area).}

\item{mufp}{minimum upstream flow path (set as one fifth of the cell size).}

\item{mascfile}{mask file}
}
\value{
List object with 2 elements:
 \describe{
  \item{hires}{RasterLayer with location of cells outlet}
  \item{lowres}{RasterStack with flow direction, outlet row, and outlet column.}
  }
}
\description{
Upscaling flow direction from fine- to coarse‐resolution grid using the
improved algorithm known as Cell Outlet Tracing with an Area Threshold
(COTAT+) (Paz et al. 2006).
}
\examples{
area_hres
dir_hres
upscaling_out<- cotat_plus(flowdir.fres = dir_hres,
                               draina.fres = area_hres,
                               factor = 10,
                               area.thres = 1.0,
                               mufp = 0.02)
# outlet locations (high resolution)
upscaling_out[[1]]
# flow direction, outlet row, outlet col
upscaling_out[[2]]
plot(upscaling_out[[2]])
flow_dir <- raster(upscaling_out[[2]], 1)
flow_dir
}
\references{
\insertRef{Paz2006}{urdn}

\insertRef{Saraiva2014}{urdn}

\insertRef{Sousa2017}{urdn}
}
\author{
Jonatan Tatsch

Nelson Navarrete
}
