% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urdn.R
\docType{package}
\name{urdn-package}
\alias{urdn}
\alias{urdn-package}
\title{Upscaling river drainage networks}
\description{
\pkg{urdn} allows you to upscaling river drainage networks.
}
\details{
The only function you're likely to need from \pkg{urdn} is
\code{\link{cotat_plus}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lhmet/urdn}
  \item Report bugs at \url{https://github.com/lhmet/urdn/issues}
}

}
\author{
\strong{Maintainer}: Jonatan Tatsch \email{jdtatsch@gmail.com}

Authors:
\itemize{
  \item Nelson Navarrete
}

Other contributors:
\itemize{
  \item Adriano Paz [contributor]
}

}
